Logo Change and ERD Added
I simplified the logo and added it to the readme. I may actually do it
again because it's more so a placeholder for now.

I added an Entity Relationship Diagram before I start work with
Supabase. This diagram has the core idea but it is possible that more
tables will be added in the future.