Added Tests For Journey Items
I added some tests for adding items to the database. These were
difficult to test for because supabase isn't liking the amount of writes
I'm doing for the email inserts. I hope to get this issue resolved with
SMTP settings. Still, the tests inserts items.

Also, the listUsers call will probably have to be changed as that could
cause strain, but for now Supabase apparently this how its done at the moment.