New Bike Upload Test and Other Adjustments
I created a bike upload test to the storage bucket, and will improve on it soon.
It will link in with the rest of the database. For now, It is working
with image uploads.

I made some smaller changes like adding packages needed, removing
unnecessary logs and utilising let instead of var.