Project Restructuring, Testing Incoming
I've been dealing with designing tests but mocking the google maps api
is lengthier than I expected. I will push a change soon. For now, I did
some restructuring of the google maps api and implemented atomic design
as mentioned in my earlier commits. It's a lot easier to test this way
and cleaner in general