Reformatted Tests and Added to Readme
This is a small commit with mostly self reflection. I designed my tests
for the Google Maps API, testing what may pop up for a user using my
system. It's good to point out that some testing is not available for
TDD. I explain this in my commit, which was found in my research today.
I think I'll go ahead with the designing the web app from my next commit