Restructuring, Tailwind Config, and Atomic Design
My config for many node modules needed to be changed. Tailwind has some
issues with the newest node version, and I needed to downgrade the
version as the version itself was unstable for what I was doing.

Tailwind is configured, mostly. It runs and has the files that it wants
but I'm still trying to understand it. It would be useful to implement.

The navbar is now split up as per atomic design. I changed the pages as
well. It's more readable and follows the design much better. The layout
is its own organism, and icons are their own atoms. Great. I just need
to do some tinkering and things will fall into place