Rearranged Navbar Icons and Added Expand Search
The navbar was starting to look a little disorganised so I removed some
sections. Five sections on the right is too much, I will move the rest
to the side bar later. I want the navbar to be simple and intuitive,
not cluttered and awkward.

I also made the search look nicer